// src/app/shared/models/IEmployee.ts
export interface IEmployee {
  EmployeeID: number;  
  Code: string;
    Name: string;
    FatherName: string;
    DOB: Date;
    NICNo : number;
    Religion: string;
    Qualification: string;
    Sex: string;
    City: string;
    Status: string
    ImageID : number;
    ImageName : string;
    
  }

  